+++ packages/analytics/dashboard-renderer/sandbox/pages/EditableDashboardDemo.vue
@@ -77,10 +77,6 @@ const context = computed<DashboardRendererContext>(() => ({
 }))
 
 const dashboardConfig = ref <DashboardConfig>({
-  gridSize: {
-    cols: 6,
-    rows: 7,
-  },
   tileHeight: 167,
   tiles: [
     {