+++ packages/entities/entities-plugins/src/components/free-form/shared/EnumField.vue
@@ -11,7 +11,9 @@
     v-bind="fieldAttrs"
     v-model="fieldValue"
     class="ff-enum-field"
+    :data-testid="field.path.value"
     :items="realItems"
+    :kpop-attributes="{ 'data-testid': `${field.path.value}-items` }"
   >
     <template
       v-if="fieldAttrs.labelAttributes?.info"