+++ packages/entities/entities-plugins/src/components/free-form/shared/ObjectField.vue
@@ -30,6 +30,7 @@
     <header class="ff-object-field-header">
       <KLabel
         class="ff-object-field-label"
+        :data-testid="`ff-label-${field.path.value}`"
         v-bind="{
           ...fieldAttrs,
           required: false
@@ -52,6 +53,7 @@
           v-if="collapsible && realAdded"
           appearance="tertiary"
           :class="`ff-object-field-button-${realExpanded ? 'collapse' : 'expand'}`"
+          :data-testid="`ff-object-toggle-btn-${field.path.value}`"
           icon
           @click="expanded = !realExpanded"
         >
@@ -64,8 +66,9 @@
           v-if="!fieldAttrs.required"
           appearance="tertiary"
           :class="`ff-object-field-button-${realAdded ? 'remove' : 'add'}`"
+          :data-testid="`ff-object-remove-btn-${field.path.value}`"
           icon
-          @click="added = !added"
+          @click="handleAddOrRemove"
         >
           <TrashIcon v-if="realAdded" />
           <AddIcon v-else />
@@ -129,7 +132,7 @@ const {
 }>()
 
 const { value: fieldValue, ...field } = useField(toRef(props, 'name'))
-const { getSchema } = useFormShared()
+const { getSchema, getDefault } = useFormShared()
 
 const added = defineModel<boolean>('added', { default: undefined })
 
@@ -183,6 +186,15 @@ const childFields = computed(() => {
     return aIndex - bIndex
   })
 })
+
+function handleAddOrRemove() {
+  added.value = !added.value
+  if (added.value) {
+    fieldValue!.value = getDefault(field.path!.value)
+  } else {
+    fieldValue!.value = null
+  }
+}
 
 watch(realAdded, (value) => {
   if (!collapsible) {