+++ packages/entities/entities-vaults/src/types/vault-form.ts
@@ -19,6 +19,11 @@ export interface BaseVaultFormConfig extends Omit<BaseFormConfig, 'cancelRoute'>
    * TODO: remove when support for approle option is added
    */
   hcvAppRoleMethodAvailable?: boolean
+
+  /**
+   * Show/hide cert option and corresponding fields
+   */
+  hcvCertMethodAvailable?: boolean
   /**
    * Show/hide Konnect Config Store option
    */
@@ -60,6 +65,7 @@ export enum VaultAuthMethods {
   TOKEN = 'token',
   K8S = 'kubernetes',
   APP_ROLE = 'approle',
+  CERT = 'cert',
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
@@ -111,6 +117,9 @@ export interface HCVVaultConfig {
   neg_ttl?: number
   resurrect_ttl?: number
   base64_decode?: boolean
+  cert_auth_role_name?: string
+  cert_auth_cert?: string
+  cert_auth_cert_key?: string
 }
 
 export interface AzureVaultConfig {
@@ -128,13 +137,14 @@ export interface AzureVaultConfig {
 
 export interface ConjurVaultConfig {
   endpoint_url: string
-  auth_method: 'default'
+  auth_method: 'api_key'
   login?: string
   account?: string
   api_key?: string
   ttl?: number
   neg_ttl?: number
   resurrect_ttl?: number
+  base64_decode?: boolean
 }
 
 // allow for nullish values in payload because Kong Admin API treats null as an empty value