+++ packages/analytics/analytics-chart/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.0.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@9.0.8...@kong-ui-public/analytics-chart@9.0.9) (2025-05-28)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+## [9.0.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@9.0.7...@kong-ui-public/analytics-chart@9.0.8) (2025-05-27)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
+## [9.0.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@9.0.6...@kong-ui-public/analytics-chart@9.0.7) (2025-05-27)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-chart
+
+
+
+
+
 ## [9.0.6](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-chart@9.0.5...@kong-ui-public/analytics-chart@9.0.6) (2025-05-22)